+++ src/xaps-utils.h
@@ -39,6 +39,7 @@ const char *get_real_mbox_user(struct mail_user *muser);
 
 /* This is data that is shared by all plugin users. */
 struct xaps_config {
+    pool_t pool;
     struct http_url *http_url;
     struct http_client *http_client;
     const char *user_lookup;